Fixed a bug in task.csv file loader
Sometimes task.csv file contains a value larger than the limit of Python
int, especially in INTERVAL, TIMING_CLOCK and GL_TIMER_VAL fields.
When the program tries to read such data, an error is raised complaining
that the number is larger than int.
Thus the program need to know that these fields should be read as long
rather than int.
This was solved by defining a converter for genfromtext(), as is done in
odml_av2.py